+++ pages/formula.php
@@ -109,8 +109,18 @@ if($form[0]['ingredient']){
                 	<input type="submit" name="add" id="add-btn" class="btn btn-info" value="Add" /> </td>  
                 </div>  
             </div>
+            
+            <div class="col-sm-6 mb-1">
+        		<input type="checkbox" name="reCalcAdd" id="reCalcAdd" value="1" data-val="1" /> Adjust solvent<i class="fa-solid fa-circle-info ml-2 pv_point_gen" rel="tip" data-placement="right" data-title="The added ingredient's quantity will be deducted from the selected solvent."></i>
+            </div>
+            <div id="slvMetaAdd">
+            	<div class="col-sm-6 mr-2 mb-1">
+        			<input name="formulaSolventsAdd" id="formulaSolventsAdd" type="text" class="formulaSolventsAdd pv-form-control">
+            	</div>
           </div>
-
+		  <div class="col-sm dropdown-divider"></div>
+        </div>
+          
           <div id="fetch_formula">
           	<div class="loader-center">
             	<div class="loader"></div>
@@ -217,19 +227,21 @@ $('#add_ing').on('click', '[id*=add-btn]', function () {
 		type: 'POST',
 		data: {
 			action: "addIng",
-			fid: "<?=$fid?>",
+			fid: myFID,
 			quantity: $("#quantity").val(),
 			concentration: $("#concentration").val(),
 			ingredient: $("#ingredient").val(),
-			dilutant: $("#dilutant").val()
+			dilutant: $("#dilutant").val(),			
+			reCalc: $("#reCalcAdd").prop('checked'),
+			formulaSolventID: $("#formulaSolventsAdd").val()
 			},
 		dataType: 'json',
 		success: function (data) {
 			if ( data.success ) {
-				var msg = '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>' + data.success + '</div>';
+				msg = '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>' + data.success + '</div>';
 				reload_formula_data();
 			} else {
-				var msg = '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>' + data.error + '</strong></div>';
+				msg = '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>' + data.error + '</strong></div>';
 			}
 			$('#msgInfo').html(msg);
 		}